skin-disease-classifier-local/requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
opencv-python-headless==4.10.0.84
Pillow==10.3.0
tqdm==4.66.5
scikit-learn==1.5.2
tensorflow==2.17.0
gunicorn==23.0.0
